+++ app/js/pagelayout.js
@@ -70,10 +70,15 @@ const hide = (view, close = false) => {
                 !== view.getAttribute("link-id"))
         }
         if (close) {
-            document.getElementById("tabs").removeChild(TABS.currentTab())
-            document.getElementById("pages").removeChild(TABS.currentPage())
+            document.getElementById("tabs").removeChild(
+                TABS.tabOrPageMatching(view))
+            document.getElementById("pages").removeChild(view)
         }
         TABS.switchToTab(TABS.listTabs().indexOf(newTab))
+    } else if (close) {
+        document.getElementById("tabs").removeChild(
+            TABS.tabOrPageMatching(view))
+        document.getElementById("pages").removeChild(view)
     }
     applyLayout()
 }
@@ -351,6 +356,7 @@ const applyLayout = () => {
 }
 
 module.exports = {
+    layoutDivById,
     switchView,
     hide,
     add,